Cleaned up mess of new files in backend
